Revert to Xcode 13.4.1
Xcode 14.0.1+ is not working in CI. What I'm observing is that
without a fastlane match --readonly, the build just fails for
not finding a certificate. With fastlane match --readonly, the build
hangs forever attempting to codesign TeakResources.

Xcode 13.4.1 does not appear to be attempting to sign TeakResources.
That's probably the difference that's causing failures. At this time
I don't know why Xcode 14+ is trying to sign TeakResources.